more int to long

+++ IndustryThing/Misc/Extensions.cs
@@ -17,7 +17,7 @@ namespace IndustryThing.Misc
 
         internal static long FindItem(this List<Asset> assets, int typeID)
         {
-            int itemsCount = 0;
+            long itemsCount = 0;
 
             foreach (var asset in assets)
             {
